requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.40
botocore==1.40.40
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dbt==1.0.0.40.5
deprecation==2.1.0
distlib==0.4.0
filelock==3.19.1
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
platformdirs==4.3.8
postgrest==1.1.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
realtime==2.7.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
websockets==15.0.1
